Refactor, clean up, add comments on intentions
